Cleanup for the package
- move test files close to source files
- stop using bot
- remove NullClient
- remove interface
