--- .envrc
@@ -1,2 +1,0 @@
-export GOPATH=$PWD/vendor:$(cd ../../../.. ; pwd)
-export PATH=$PWD/vendor/bin:$(cd ../../../../bin ; pwd):$PATH
