+++ .travis.yml
@@ -1,7 +1,18 @@
 language: go
+sudo: false
 
 go:
   - 1.3
   - 1.4
+  - 1.5
+  - tip
 
-script: cd tests && go test
+matrix:
+  allow_failures:
+    - go: tip
+
+install:
+  - go get -v -t ./...
+
+script:
+  - go test -race ./...
